refactor: move cell components into dedicated folders
