+++ src/components/DataTable/cells/TableCell/TableCell.module.scss
@@ -1,4 +1,4 @@
-@use '../../../utils/index.scss';
+@use '../../../../utils/index.scss';
 
 .tableCell {
     gap: var(--spacing-nano);
