+++ src/components/DataTable/cells/TableCellActions/TableCellActions.types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import type {TableCellProps} from '../TableCell.types';
+import type {TableCellProps} from '../TableCell/TableCell.types';
 
 /**
  * Props for TableCellActions - a cell dedicated to row actions.
