+++ src/components/DataTable/cells/TableCellStatus/TableCellStatus.types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import type {TableCellProps} from '../TableCell.types';
+import type {TableCellProps} from '../TableCell/TableCell.types';
 
 export type TableCellStatusProps = Omit<TableCellProps, 'color' | 'children'> & {
     /**
