+++ src/components/DataTable/cells/TableHeadCell/TableHeadCell.types.ts
@@ -1,5 +1,5 @@
 import React from 'react';
-import type {TableCellProps} from './TableCell.types';
+import type {TableCellProps} from '../TableCell/TableCell.types';
 
 export type TableHeadCellProps = Omit<TableCellProps, 'isScrollable' | 'component'> & {
     /**
