+++ src/components/DataTable/cells/TableStructuredCell/TableStructuredCell.module.scss
