+++ src/components/DataTable/cells/TableStructuredCell/TableStructuredCell.spec.tsx
